requirements.txt
fastapi
twilio
uvicorn
requests
python-dotenv
openai
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
ultravox-client